requirements.txt
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.2
bcrypt==3.2.0
blessed==1.19.0
blis==0.7.4
botocore==1.21.54
bugsnag==4.1.0
cached-property==1.5.2
catalogue==2.0.4
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.1
click==7.1.2
cloudinary==1.26.0
colorama==0.4.3
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cymem==2.0.5
dj-database-url==0.5.0
Django==3.2.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-environ==0.4.5
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djangorestframework-stubs==1.4.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
flake8==3.9.2
future==0.16.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
leoprofanity==0.0.4
mailjet-rest==1.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
more-itertools==8.8.0
murmurhash==1.0.5
mypy==0.910
mypy-extensions==0.4.3
now==0.0.6
numpy==1.21.0
ordered-set==4.0.2
ordered-set-stubs==0.1.3
packaging==21.0
paramiko==2.7.2
pathspec==0.5.9
pathy==0.6.0
pip-upgrader==1.4.15
plac==1.3.3
poetry-version==0.1.5
preshed==3.0.5
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.0
semantic-version==2.8.5
sendgrid==6.7.1
sib-api-v3-sdk==7.2.0
six==1.14.0
smart-open==5.1.0
spacy==3.1.0
spacy-legacy==3.0.8
sqlparse==0.4.1
srsly==2.4.1
starkbank-ecdsa==1.1.1
termcolor==1.1.0
terminaltables==3.1.0
texttable==1.6.4
thinc==8.0.7
threadpoolctl==2.2.0
toml==0.10.2
tomlkit==0.5.11
tqdm==4.61.2
typer==0.3.2
types-pytz==2021.1.2
types-PyYAML==5.4.10
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.1.9
WebOb==1.8.7
websocket-client==0.59.0
whitenoise==5.2.0
